Follow links in stack traces to line and file in PhpStorm
After installing `webapp` package we have access to Symfony Profiler at the bottom of the page at http://localhost/.

This commit makes it possible to go directly to a line and file in PhpStorm by clicking on the filenames that Xdebug
shows in stack traces in Symfony Profiler.

Do `make restart` after adding this.

For this to work PhpStorm plugin *IDE Remote Control* is needed:

- https://youtrack.jetbrains.com/issue/IDEA-308020
- https://plugins.jetbrains.com/plugin/19991-ide-remote-control/reviews